First PIR Sensor Demo based on Raspberry Pi
There are two demo widgets in this project.

PIRDemoWidget is a plot widget based on QCustomPlot, which shows history
data and current value, and visualizes current value with color.

PIRDemoSimpleWidget is a simplified version of former, which has an
indicator shows the occupancy status, also raw data from sensor, and a
color visualization. But it is not using QCustomPlot thus compiles
faster.